Moved nth to List data type
